Metrics extension is disabled if system property `ts.prometheus-http-endpoint` is not defined
